test(api): break out tools for fuzz testing type validators (#134)
* test(api): break out tools for fuzz testing type validators

* test(api): break out tools for fuzz testing type validators

* refactor based on PR (#134)
