+++ libs/data-types/src/index.ts
@@ -1,4 +1,5 @@
 export * from './lib/decorators';
 export * from './lib/enums';
+export * from './lib/fuzz';
 export * from './lib/types';
 export * from './lib/utilities';
