+++ libs/data-types/src/lib/exceptions/FailedToGenerateFuzzForUnsupportedDataTypeException.ts
@@ -1,0 +1,5 @@
+export class FailedToGenerateFuzzForUnsupportedDataTypeException extends Error {
+    constructor(type: string) {
+        super(`Failed to generate fuzz for unsupported data type: ${type}`);
+    }
+}
