+++ libs/data-types/src/lib/fuzz/FuzzGenerator.ts
@@ -1,0 +1,10 @@
+import { CoscradDataSchema } from '../types/CoscradDataSchema';
+import generateInvalidValuesForProperty from './generateInvalidValuesForProperty';
+
+export class FuzzGenerator {
+    constructor(private readonly schema: CoscradDataSchema) {}
+
+    generateInvalidValues() {
+        return generateInvalidValuesForProperty(this.schema);
+    }
+}
