+++ libs/data-types/src/lib/fuzz/index.ts
@@ -1,0 +1,3 @@
+import { FuzzGenerator } from './FuzzGenerator';
+
+export { FuzzGenerator };
