+++ libs/data-types/src/lib/types/CoscradDataSchema.ts
@@ -1,0 +1,7 @@
+import { CoscradDataType } from './CoscradDataType';
+
+export type CoscradDataSchema = {
+    type: CoscradDataType;
+    isArray: boolean;
+    isOptional: boolean;
+};
