+++ libs/data-types/src/lib/types/index.ts
@@ -1,3 +1,4 @@
 export * from './ClassDataTypeMetadata';
+export * from './CoscradDataSchema';
 export * from './CoscradDataType';
 export * from './PropertyTypeMetadata';
